Final commit
Now when the ball touches one of the screen sides, it gets the color of
the player who scored the point, the whole game pauses for 1 second and
then the ball color and position are reset and the velocities randomized
again.

Timing problem fixed - now the game can run continuously on the same
speed, after local corrections in the TIMER module and adaptation on
the sleep() function (all variables sizes on iob-timer.c and .h were
changed to uint64_t to handle all possible values instead of various
conversions with losses of data).

Reset button changed to reset switch because the issue with the black
screen appears to be caused by pressing the button for some reason. Also
corrected the implementation of the register (now it has the correct
size - 1 byte).

Added the output of the data of the controller 1 on NESCTRL (just in
case it's useful).

And removed all LED pins usages (removed the controllers data outputs
usages) to reduce the pin usage number to the minimum and actually
useful number.

+++ software/firmware/Utils.c
@@ -1,7 +1,7 @@
 #include "iob-timer.h"
 
-void sleep(unsigned ms) {
-	unsigned init_time = timer_time_ms();
+void sleep(uint64_t ms) {
+	uint64_t init_time = timer_time_ms();
 	while (timer_time_ms() - init_time < ms) {
 		// Do nothing
 	}
